style: normに対応
